backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.33.0
sqlalchemy==2.0.43
psycopg[binary]==3.1.18
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
redis==6.1.1
celery==5.5.3
httpx==0.28.1
python-multipart==0.0.20
email-validator==2.2.0
pytest==8.3.5
pytest-asyncio==0.24.0
